backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
aiofiles==24.1.0
alembic==1.14.1
pyautogui==0.9.54
Pillow==11.1.0
uiautomation==2.0.29
python-docx==1.2.0
PyPDF2==3.0.1
openpyxl==3.1.5
